requirements.txt
matrix-nio
python-vlc
